pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "commandclaw"
version = "0.1.0"
description = "Git-native AI agent platform — vault is the control plane"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "Shikhar", email = "shikhar@commandclaw.dev" }]

dependencies = [
    "langchain>=0.3,<0.4",
    "langchain-openai>=0.3,<0.4",
    "python-telegram-bot>=21,<22",
    "langfuse>=2,<3",
    "gitpython>=3.1,<4",
    "pydantic>=2,<3",
    "pydantic-settings>=2,<3",
    "python-frontmatter>=1,<2",
    "python-dotenv>=1,<2",
    "mcp>=1,<2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.24,<1",
    "pytest-cov>=5,<6",
    "ruff>=0.8,<1",
]

[project.scripts]
commandclaw = "commandclaw.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/commandclaw"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
